// src/integrations/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
// ...
      menu: {
        Row: {
          breakfast: string | null
          created_at: string
          day: string
          dinner: string | null
          id: string
          lunch: string | null
          optional_dishes: Json | null
          owner_id: string
          updated_at: string
          week_number: number
        }
        Insert: {
          breakfast?: string | null
          created_at?: string
          day: string
          dinner?: string | null
          id?: string
          lunch?: string | null
          optional_dishes?: Json | null
          owner_id: string
          updated_at?: string
          week_number?: number
        }
        Update: {
          breakfast?: string | null
          created_at?: string
          day?: string
          dinner?: string | null
          id?: string
          lunch?: string | null
          optional_dishes?: Json | null
          owner_id?: string
          updated_at?: string
          week_number?: number
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}